lots of changes made

+++ BPAddin/BPAddin/Import/Importer.cs
@@ -1,4 +1,6 @@
 ﻿using System.IO;
+using System.Windows.Forms;
+using static BPAddin.util.EAMacros;
 
 namespace BPAddin
 {
@@ -8,20 +10,12 @@ namespace BPAddin
         private EA.Package uiLibPkg;
         private EA.Diagram uiDiagram;
 
-        private const string uiLibName = "UI Library";
-        private const string uiLibPkgType = "Package";
-        private const string uiLibDiaType = "Class"; 
-        private const string UI_EL_STEREOTYPE = "UIElement";
-
-        private string uiLibPath;
-
         public Importer(EA.Repository repository) { 
             this.repository = repository;
         }
 
 
-        public void importUILibrary() {
-            uiLibPath = Properties.Settings.Default.UILibraryDir;
+        public void importUILibrary(string uiLibPath) {
             
             EA.Project project = repository.GetProjectInterface();
             string pkgGUID = project.GUIDtoXML(uiLibPkg.PackageGUID);
@@ -37,22 +31,25 @@ namespace BPAddin
                 
                 project.ImportFile(pkgGUID, "C#", file, "");
             }
-
             
-            setStereotypesForClasses(uiLibPkg, UI_EL_STEREOTYPE);
+            setStereotypesForClasses(uiLibPkg, STYPE_UIELEMENT);
         }
 
-        public void initUILibrary()
+        public void initUILibrary(string uiLibPath)
         {
-            uiLibPkg = findOrCreatePackage(uiLibName);
-            uiDiagram = findOrCreateDiagram(uiLibPkg, uiLibName);
+            uiLibPkg = findOrCreatePackage(UILIB_PKGNAME);
+            if(uiLibPkg == null)
+                return;
 
-            importUILibrary();
+            uiDiagram = findOrCreateDiagram(uiLibPkg, UILIB_PKGNAME);
+
+            importUILibrary(uiLibPath);
             populateDiagram(uiLibPkg, uiDiagram);
             repository.RefreshModelView(0);
         }
 
         private EA.Package findOrCreatePackage(string pkgName) {
+            // find the package based on the root
             EA.Package root = (EA.Package)repository.Models.GetAt(0);
 
             // find package in root
@@ -62,17 +59,25 @@ namespace BPAddin
 
                 if(pkg.Name == pkgName)
                 {
-                    //MessageBox.Show("Package found: " + pkgName);
                     // delete existing package
+                    if (MessageBox.Show(
+                        "An existing UI Library has been found.\n" +
+                        "Importing now will replace existing UI Library.\n" +
+                        "\nContinue?",
+                        "BPAddin", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) 
+                        != DialogResult.Yes)
+                        return null;
+
                     root.Packages.DeleteAt(i, true);
                     root.Packages.Refresh();
+
                     break;
                 }
             }
 
             // create package to root
             //MessageBox.Show("Creating new package - " + pkgName + " as type " + uiLibPkgType);
-            EA.Package pkgNew = (EA.Package)root.Packages.AddNew(uiLibName, uiLibPkgType);
+            EA.Package pkgNew = (EA.Package)root.Packages.AddNew(UILIB_PKGNAME, UILIB_PKGTYPE);
             pkgNew.Update();
 
             root.Packages.Refresh();
@@ -84,14 +89,14 @@ namespace BPAddin
             // find diagram
             foreach(EA.Diagram diagram in pkg.Diagrams)
             {
-                if(diagram.Type == uiLibDiaType && diagram.Name == name)
+                if(diagram.Type == UILIB_DIAGRAMTYPE && diagram.Name == name)
                 {
                     return diagram;
                 }
             }
 
             // create diagram
-            EA.Diagram diagramNew = (EA.Diagram)pkg.Diagrams.AddNew(uiLibName, uiLibDiaType);
+            EA.Diagram diagramNew = (EA.Diagram)pkg.Diagrams.AddNew(UILIB_PKGNAME, UILIB_DIAGRAMTYPE);
             diagramNew.Update();
             pkg.Diagrams.Refresh();
             
@@ -103,7 +108,7 @@ namespace BPAddin
             int count = 0;
             foreach (EA.Element el in pkg.Elements)
             {
-                if (el.Type == "Class")
+                if (el.Type == ELTYPE_CLASS)
                 {
                     el.Stereotype = stereotype;
                     el.Update();
@@ -121,7 +126,7 @@ namespace BPAddin
         {
             foreach(EA.Element el in pkg.Elements)
             {
-                if (el.Type != "Class")
+                if (el.Type != ELTYPE_CLASS)
                     continue;
 
                 EA.DiagramObject dobj = (EA.DiagramObject)diagram.DiagramObjects.AddNew(el.Name, el.Type);
